+++ src/RcsbFv3D/RcsbFv3DSequenceIdentity.tsx
@@ -37,8 +37,8 @@ import {
     AlignmentTrajectoryParamsType
 } from "../RcsbFvStructure/StructureViewers/MolstarViewer/TrajectoryPresetProvider/AlignmentTrajectoryPresetProvider";
 import {
-    MolstarComponentActionFactory
-} from "../RcsbFvStructure/StructureViewers/MolstarViewer/MolstarUtils/MolstarComponentAction";
+    MolstarAlignmentComponentActionFactory
+} from "../RcsbFvStructure/StructureViewers/MolstarViewer/MolstarUtils/MolstarAlignmentComponentAction";
 import {MolstarTools} from "../RcsbFvStructure/StructureViewers/MolstarViewer/MolstarUtils/MolstarTools";
 import getModelIdFromTrajectory = MolstarTools.getModelIdFromTrajectory;
 
@@ -102,7 +102,7 @@ export class RcsbFv3DSequenceIdentity extends RcsbFv3DAbstract<
             >( new MolstarManagerFactory(getModelIdFromTrajectory) ),
             structureViewerBehaviourObserver: new MsaBehaviourObserver<AlignmentLoadMolstarType,LoadMolstarReturnType>(
                 new MolstarAlignmentLoader(),
-                new MolstarComponentActionFactory()
+                new MolstarAlignmentComponentActionFactory()
             )
         });
     }